feat: 읽지않은 수 실시간 반영 로직 추가
- 읽지않은 수 파악을 위해 redis에 읽지않은 참가자 수 저장하도록 redis 추가
- 스케쥴러로 redis에 저장되어있는 데이터를 db에 주기적으로 백업하도록 구현
- 실시간으로 읽지않은숫자를 메시지에 반영할 수 있도록 로직 구현

+++ src/main/java/org/example/paymenttest/entity/chat/OpenChatRoom.java
@@ -20,5 +20,5 @@ public class OpenChatRoom {
     private String imageUrl;
     private String lastMessage;
     private LocalDateTime lastTimestamp;
-    private int participantCount;
+    private Long participantCount;
 }
